+++ jest.config.js
@@ -2,7 +2,6 @@ const config = require('@polkadot/dev/config/jest');
 
 module.exports = Object.assign({}, config, {
   moduleNameMapper: {
-    '@polkadot/api-(observable)(.*)$': '<rootDir>/packages/api-$1/src/$2',
     '@polkadot/api$': '<rootDir>/packages/api/src/$1',
     '@polkadot/rpc-(core|provider|rx)(.*)$': '<rootDir>/packages/rpc-$1/src/$2',
     '@polkadot/extrinsics(.*)$': '<rootDir>/packages/type-extrinsics/src/$1',
@@ -12,7 +11,6 @@ module.exports = Object.assign({}, config, {
   },
   modulePathIgnorePatterns: [
     '<rootDir>/packages/api/build',
-    '<rootDir>/packages/api-observable/build',
     '<rootDir>/packages/rpc-core/build',
     '<rootDir>/packages/rpc-provider/build',
     '<rootDir>/packages/rpc-rx/build',