+++ packages/api/src/promise/types.ts
@@ -1,4 +1,4 @@
-// Copyright 2017-2018 @polkadot/api authors & contributors
+// Copyright 2017-2019 @polkadot/api authors & contributors
 // This software may be modified and distributed under the terms
 // of the Apache-2.0 license. See the LICENSE file for details.
 
@@ -8,11 +8,31 @@ import { MethodFunction } from '@polkadot/types/Method';
 import { StorageFunction } from '@polkadot/types/StorageKey';
 import { ApiBaseInterface } from '../types';
 
-import Rpc from '@polkadot/rpc-core/index';
-
 import SubmittableExtrinsic from './SubmittableExtrinsic';
 
+export type UnsubFunction = () => void;
+
+export type DecoratedRpc$Method = (...params: Array<any>) => Promise<any> | UnsubFunction;
+
+export interface DecoratedRpc$Section {
+  [index: string]: DecoratedRpc$Method;
+}
+
+export interface DecoratedRpc {
+  author: DecoratedRpc$Section;
+  chain: DecoratedRpc$Section;
+  state: DecoratedRpc$Section;
+  system: DecoratedRpc$Section;
+}
+
+export interface QueryableStorageFunction$Subscribe {
+  (arg: any, cb: (value?: any | null) => any): UnsubFunction;
+  (cb: (value?: any | null) => any): UnsubFunction;
+}
+
 export interface QueryableStorageFunction extends StorageFunction {
+  (cb: (value?: any | null) => any): UnsubFunction;
+  (arg: any, cb: (value?: any | null) => any): UnsubFunction;
   (arg?: any): Promise<Codec | null | undefined>;
   at: (hash: Hash, arg?: any) => Promise<Codec | null | undefined>;
 }
@@ -37,7 +57,7 @@ export interface SubmittableExtrinsics {
   [index: string]: SubmittableModuleExtrinsics;
 }
 
-export interface ApiPromiseInterface extends ApiBaseInterface<Rpc, QueryableStorage, SubmittableExtrinsics> {
+export interface ApiPromiseInterface extends ApiBaseInterface<DecoratedRpc, QueryableStorage, SubmittableExtrinsics> {
   readonly isReady: Promise<ApiPromiseInterface>;
 }