+++ packages/rpc-provider/src/ws/index.ts
@@ -1,4 +1,4 @@
-// Copyright 2017-2018 @polkadot/rpc-provider authors & contributors
+// Copyright 2017-2019 @polkadot/rpc-provider authors & contributors
 // This software may be modified and distributed under the terms
 // of the Apache-2.0 license. See the LICENSE file for details.
 
@@ -330,6 +330,12 @@ export default class WsProvider implements WSProviderInterface {
 
     Object.keys(subscriptions).forEach(async (id) => {
       const { callback, method, params, type } = subscriptions[id];
+
+      // only re-create methods with '<section>_subscribe`, i.e. submissions such
+      // as 'author_submitAndWatchExtrinsic' are not included
+      if (method.indexOf('_subscribe') === -1) {
+        return;
+      }
 
       try {
         await this.subscribe(type, method, params, callback);