+++ packages/type-jsonrpc/src/system.ts
@@ -1,4 +1,4 @@
-// Copyright 2017-2018 @polkadot/jsonrpc authors & contributors
+// Copyright 2017-2019 @polkadot/jsonrpc authors & contributors
 // This software may be modified and distributed under the terms
 // of the Apache-2.0 license. See the LICENSE file for details.
 
@@ -32,6 +32,12 @@ const properties: RpcMethodOpt = {
   type: 'ChainProperties'
 };
 
+const health: RpcMethodOpt = {
+  description: 'Return health status of the node',
+  params: [],
+  type: 'Health'
+};
+
 const section = 'system';
 
 /**
@@ -44,6 +50,7 @@ export default {
   section,
   methods: {
     chain: createMethod(section, 'chain', chain),
+    health: createMethod(section, 'health', health),
     name: createMethod(section, 'name', name),
     properties: createMethod(section, 'properties', properties),
     version: createMethod(section, 'version', version)