+++ packages/types/src/SignedBlock.ts
@@ -1,4 +1,4 @@
-// Copyright 2017-2018 @polkadot/types authors & contributors
+// Copyright 2017-2019 @polkadot/types authors & contributors
 // This software may be modified and distributed under the terms
 // of the Apache-2.0 license. See the LICENSE file for details.
 
@@ -6,6 +6,7 @@ import { AnyU8a } from './types';
 
 import Struct from './codec/Struct';
 import Block, { BlockValue } from './Block';
+import Hash from './Hash';
 import Justification from './Justification';
 
 type SignedBlockValue = {
@@ -31,6 +32,13 @@ export default class SignedBlock extends Struct {
    */
   get block (): Block {
     return this.get('block') as Block;
+  }
+
+  /**
+   * @description Block/header [[Hash]]
+   */
+  get hash (): Hash {
+    return this.block.hash;
   }
 
   /**