+++ packages/types/src/Type.ts
@@ -1,4 +1,4 @@
-// Copyright 2017-2018 @polkadot/types authors & contributors
+// Copyright 2017-2019 @polkadot/types authors & contributors
 // This software may be modified and distributed under the terms
 // of the Apache-2.0 license. See the LICENSE file for details.
 
@@ -56,9 +56,9 @@ export default class Type extends Text {
       // alias Vec<u8> -> Bytes
       Type._alias('Vec<u8>', 'Bytes'),
       // alias RawAddress -> Address
-      Type._alias('RawAddress', 'Address')
-      // TODO Check these for possibly matching -
-      //   `PropIndex` -> `ProposalIndex` (implementation looks the same, however meant as diff)
+      Type._alias('RawAddress', 'Address'),
+      // flattens tuples with one value, `(AccountId)` -> `AccountId`
+      Type._flattenSingleTuple()
     ];
 
     return mappings.reduce((result, fn) => {
@@ -129,6 +129,12 @@ export default class Type extends Text {
       }
 
       return value;
+    };
+  }
+
+  private static _flattenSingleTuple (): Mapper {
+    return (value: string): string => {
+      return value.replace(/\(([^,]*)\)/, '$1');
     };
   }